# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This a progress management system API.",
        "title": "JailBird API",
        "contact": {
            "name": "API support and issue management",
            "url": "https://electratype.com",
            "email": "support@electratype.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5454",
    "basePath": "/api/v1",
    "paths": {
        "/projects": {
            "get": {
                "description": "Returns array of all projects",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "Return all projects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Project"
                            }
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "project"
                ],
                "summary": "Create new project",
                "parameters": [
                    {
                        "description": "Project definition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PlainProject"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/projects/{id}": {
            "delete": {
                "tags": [
                    "project"
                ],
                "summary": "Delete project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "main.PlainProject": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "description": {
                    "description": "Description of the project\nRequired: false",
                    "type": "string"
                },
                "id": {
                    "description": "The id/slug for the project\nRequired: true\nMin length: 1",
                    "type": "string",
                    "minLength": 1
                },
                "name": {
                    "description": "Name of the project\nRequired: false",
                    "type": "string"
                }
            }
        },
        "main.Project": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "description": "Description of the project\nRequired: false",
                    "type": "string"
                },
                "id": {
                    "description": "The id/slug for the project\nRequired: true\nMin length: 1",
                    "type": "string",
                    "minLength": 1
                },
                "name": {
                    "description": "Name of the project\nRequired: false",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        }
    }
}